refactor: 컴포넌트 경로 수정 및 불필요한 상태 제거
