+++ src/app/check-camera/@modal/(.)finish-check/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import FinishCheckModal from '../../components/FinishCheckModal';
 
 const Modal = () => {
