+++ src/app/check-voice/sections/VoiceSection.tsx
@@ -6,7 +6,7 @@ import dynamic from 'next/dynamic';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 import { useVoiceVisualizer } from 'react-voice-visualizer';
-import Checked from '@/assets/icons/checked.svg';
+import Checked from '@assets/icons/checked.svg';
 
 const VoiceVisualizer = dynamic(
   () => import('react-voice-visualizer').then((mod) => mod.VoiceVisualizer),
