+++ src/app/home/@modal/(.)progress/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import ProgressModal from '../../components/ProgressModal';
 
 export default function Modal() {
