+++ src/app/home/sections/index.ts
@@ -1,2 +1,2 @@
-export * from './InfoSection';
-export * from './LogSection';
+export { default as InfoSection } from './InfoSection';
+export { default as LogSection } from './LogSection';
