+++ src/app/interview/@modal/(.)end-interview/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import CompleteInterviewModal from '../../components/CompleteInterviewModal';
 
 const EndModal = () => {
