+++ src/app/interview/@modal/(.)quit-interview/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import QuitInterviewModal from '../../components/QuitInterviewModal';
 
 const QuitModal = () => {
