+++ src/app/interview/page.tsx
@@ -6,20 +6,30 @@ import { useEffect, Suspense, lazy } from 'react';
 import { useInterview } from '@/hooks/useInterview';
 import Loading from '../loading';
 import { useInterviewQuestions } from './hooks/useInterviewQuestion';
+import { useRouter } from 'next/navigation';
+import { PATH } from '@/constants';
+import { getStorageItem } from '@/utils/storage';
 
 const QuestionBoxSection = lazy(() => import('./sections/QuestionBoxSection'));
 const VideoSection = lazy(() => import('./sections/VideoSection'));
 const SideSection = lazy(() => import('./sections/SideSection'));
 
 const Interview = () => {
+  const router = useRouter();
   const { data, isLoading } = useInterviewQuestions();
   const { resetInterview } = useInterview();
 
   useEffect(() => {
+    const interviewId = getStorageItem('interviewId');
+    if (!interviewId) {
+      router.replace(PATH.HOME);
+      return;
+    }
     resetInterview();
-  }, []);
+  }, [router, resetInterview]);
 
   if (isLoading) return <Loading />;
+  if (!data) return null;
 
   return (
     <PageLayout
