+++ src/app/modify-resume/@modal/(.)modify-confirm/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <></>;
+}
