+++ src/app/modify-resume/@modal/(.)modify-confirm/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { ModifyConfirmModal } from '../../components';
 
 const ConfirmModal = () => {
