+++ src/app/modify-resume/sections/TitleSection.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Button from '@/components/Button';
 
 export default function TitleSection() {
